some edit

+++ Data-Structures/linkedList/app/src/main/java/linkedList/Node.java
@@ -3,6 +3,5 @@ package linkedList;
 public class Node {
 public int value;
 public Node next;
-public Node(int value){this.value = (int) value;}
-
+public Node(int value){this.value = value;}
 }
